+++ packages/shopify-app-kit/tsup.config.ts
@@ -16,5 +16,5 @@ export default defineConfig({
   dts: true,
   sourcemap: true,
   clean: true,
-  external: ["react", "react-dom", "react/jsx-runtime"]
+  external: ["@shopify/polaris", "react", "react-dom", "react/jsx-runtime"]
 });